Next version of project infra

+++ Makefile
@@ -4,12 +4,13 @@ $(shell cp -n \Makefile.local.default \Makefile.local)
 $(shell cp -n \.\/src\/docker\/docker-compose\.override\.yml\.default \.\/src\/docker\/docker-compose\.override\.yml)
 include .env
 
-# Include some usefull goals.
-include Makefile.tools
+all: | net build install info
 
-all: | include net build install info
+# Include some usefull goals.
+-include Makefile.tools
 
 # First we need to check if user configured project.
+
 include:
 ifeq ($(strip $(COMPOSE_PROJECT_NAME)),projectname)
 #todo: ask user to make a project name and mv folders.
@@ -18,9 +19,14 @@ endif
 
 
 # Totally clear containers and data inside.
+  RUNNING_CONTAINERS := $(shell docker ps -f name=$(COMPOSE_PROJECT_NAME) --format "{{ .ID }}")
+
 clean: info
+	@echo $(RUNNING_CONTAINERS)
+ifeq ($(strip $(RUNNING_CONTAINERS)),)
+	@echo "Nothing to remove for $(COMPOSE_PROJECT_NAME)"
+else
 	@echo "Removing networks for $(COMPOSE_PROJECT_NAME)"
-ifeq ($(shell docker inspect --format="{{ .State.Running }}" $(COMPOSE_PROJECT_NAME)_php 2> /dev/null),true)
 	docker-compose down
 endif
 	if [ -d "build" ]; then docker run --rm -v $(shell pwd):/mnt skilldlabs/$(PHP_IMAGE) ash -c "rm -rf /mnt/build"; fi
@@ -31,16 +37,20 @@ build: clean
 	cp src/composer/docroot/index.php build/docroot
 	mkdir -p /dev/shm/${COMPOSE_PROJECT_NAME}_mysql
 
+# Installing project.
 install:
 	@echo "Updating containers..."
 	docker-compose pull
 	@echo "Build and run containers..."
 	docker-compose up -d
 	docker-compose exec -T php apk add --no-cache git shadow
+	docker-compose exec -T php apk add --no-cache git shadow
+	docker-compose exec -T php apk add --no-cache inotify-tools
 	make -s build_users
 	docker-compose exec -T php chown -hR $(UID):$(GID) /var/www/html
 	make -s reinstall
 
+# Start Drupal installation.
 reinstall:
 	cp src/composer/composer.json build/;
 ifeq ($(shell test -e src/composer/composer.lock && echo -n yes),yes)
@@ -48,7 +58,10 @@ ifeq ($(shell test -e src/composer/composer.lock && echo -n yes),yes)
 endif
 	make -s composermerge;
 	$(call execute,php,composer install --prefer-dist --optimize-autoloader);
+ifneq ($(shell test -e build/sites/default/settings.php && echo -n yes),yes)
 	$(call execute,php,cp sites/default/default.settings.php sites/default/settings.php);
+	$(call execute,php,sed -i "s/.*file_public_path.*/\$$settings['file_public_path'] =  'docroot\/files';/" sites/default/settings.php);
+endif
 
 ifeq ($(shell test -e site_settings/settings.$(ENV).php && echo -n yes),yes)
 	@echo "We have settings for $(ENV) environment, including it as settings.local.php";
@@ -56,11 +69,10 @@ ifeq ($(shell test -e site_settings/settings.$(ENV).php && echo -n yes),yes)
 	cp site_settings/settings.$(ENV).php build/settings.local.php; \
 	$(call execute,php,mv settings.local.php sites/default/settings.local.php);
 endif
-	$(call execute,php,sed -i "s/.*file_public_path.*/\$$settings['file_public_path'] =  'docroot\/files';/" sites/default/settings.php);
 	make -s si;
 	$(call execute,php,sh -c "cd docroot && ln -s ../robots.txt && ln -s ../.htaccess");
 
-
+# Merge several composer.json files.
 composermerge:
 ifeq ($(shell test -e src/composer/composer.$(ENV).json && echo -n yes),yes)
 	@echo "We have local json for $(ENV) environment, including it";
@@ -72,6 +84,8 @@ ifeq ($(shell test -e src/composer/composer.$(ENV).json && echo -n yes),yes)
 	$(call execute,php,composer config extra.merge-plugin.merge-extra true);
 	$(call execute,php,composer config extra.merge-plugin.merge-extra-deep true);
 	$(call execute,php,composer config extra.merge-plugin.merge-scripts true);
+else
+	$(info "Nothing to merge")
 endif
 
 
@@ -81,29 +95,38 @@ update:
 	$(call execute,php,drush updb -y);
 	$(call execute,php,drush pmu $(MODULES) -y);
 	$(call execute,php,drush en $(MODULES) -y);
-	make -s trans; \
 	make -s front; \
-	make -s postinstall; \
 	make -s csim; \
-	make -s rsync; \
-	make -s info
-
-si:
-	@echo "Installing from: $(PROJECT_INSTALL)"
-ifeq ($(PROJECT_INSTALL), config)
-	$(call execute,php,drush si config_installer --db-url=mysql://d8:d8@mysql/d8 --account-pass=admin -y config_installer_sync_configure_form.sync_directory=config/sync);
-	$(call execute,php,drush eval '\Drupal::service("entity.definition_update_manager")->applyUpdates();');
-	$(call execute,php,drush pmu $(MODULES) -y);
-else
-	docker-compose exec -T php drush si $(PROFILE_NAME) --db-url=mysql://d8:d8@mysql/d8 --account-pass=admin -y --site-name="$(SITE_NAME)"
-endif
-	$(call execute,php,drush en $(MODULES) -y);
 	make -s trans; \
 	make -s postinstall; \
-	make -s csim; \
 	make -s rsync; \
 	make -s info
 
+
+# Install site, in case we dont have config folder, install minimal profile and export config
+si:
+	@echo "Installing $(SITE_NAME)"
+ifeq ($($(ls -A config/sync)),)
+	$(call execute,php,drush si minimal --db-url=mysql://d8:d8@mysql/d8 --account-pass=admin -y --site-name="$(SITE_NAME)"); \
+	$(call execute,php,drush en config_split -y); \
+	docker exec $(COMPOSE_PROJECT_NAME)_php sed -i "s/.*config_directories.*/\$$config_directories['sync'] =  'config\/sync';/" sites/default/settings.php;
+	make -s csex
+else
+	$(call execute,php,drush si config_installer --db-url=mysql://d8:d8@mysql/d8 --account-pass=admin -y config_installer_sync_configure_form.sync_directory=config/sync);
+	$(call execute,php,drush entup);
+endif
+ifeq ($(shell test -d src/modules/$(MODULES) && echo -n yes),yes)
+	$(call execute,php,drush pmu $(MODULES) -y);
+	$(call execute,php,drush en $(MODULES) -y);
+endif
+	make -s front; \
+	make -s csim; \
+	make -s trans; \
+	make -s postinstall; \
+	make -s rsync; \
+	make -s info
+
+# Goal to build frontend part.
 front:
 	@echo "Building front tasks..."
 	docker pull skilldlabs/frontend:zen; \
@@ -123,7 +146,6 @@ endif
 # Get subnet ip range.
 iprange:
 	$(shell grep -q -F 'IPRANGE=' .env || echo "\nIPRANGE=$(shell docker network inspect $(COMPOSE_PROJECT_NAME)_front --format '{{(index .IPAM.Config 0).Subnet}}')" >> .env)
-
 
 
 # All tool are moved to Makefile.tools
